requirements.txt
certifi==2018.4.16
future==0.16.0
python-telegram-bot==10.0.2